change test name to AsyncPublisher and merge AsyncPublisher's Subscriber
