+++ Sources/Foo/Main.swift
@@ -1,0 +1,17 @@
+//
+//  File.swift
+//  
+//
+//  Created by 박병관 on 6/2/24.
+//
+
+import Foundation
+
+@main
+enum Main {
+    
+    static func main() throws {
+        
+    }
+    
+}
